feat(theme): Add new public themes
BREAKING CHANGE: mineralTheme, styleVariables, and color moved from ‘utils’ to ‘themes’.  Theme import syntax change.
