+++ src/themes/index.js
@@ -15,11 +15,15 @@
  */
 
 /* @flow */
-/**
-  * Helper to convert a px value to ems, relative to the base font size
-  */
-import fontSize_base from '../themes/fontSizeBase';
+export { default as mineral } from './mineral';
+export { default as blue } from './blue';
+export { default as dusk } from './dusk';
+export { default as indigo } from './indigo';
+export { default as lime } from './lime';
+export { default as purple } from './purple';
+export { default as sky } from './sky';
+export { default as slate } from './slate';
+export { default as teal } from './teal';
 
-export default function pxToEm(value: number) {
-  return `${value / fontSize_base}em`;
-}
+export { default as color } from './color';
+export { default as styleVariables } from './styleVariables';
