+++ src/themes/purple.js
@@ -15,11 +15,6 @@
  */
 
 /* @flow */
-/**
-  * Helper to convert a px value to ems, relative to the base font size
-  */
-import fontSize_base from '../themes/fontSizeBase';
+import createTheme from './createTheme';
 
-export default function pxToEm(value: number) {
-  return `${value / fontSize_base}em`;
-}
+export default createTheme('purple');
