+++ src/website/app/demos/Menu/examples/menu-item/customRender.js
@@ -18,9 +18,9 @@
 import {
   createStyledComponent,
   getNormalizedValue,
-  pxToEm,
-  mineralTheme
+  pxToEm
 } from '../../../../../../utils';
+import mineralTheme from '../../../../../../themes/mineral';
 import Menu, { MenuItem } from '../../../../../../Menu';
 import DemoLayout from '../../components/DemoLayout';
