+++ src/website/app/demos/Menu/examples/menu-item/states.js
@@ -16,7 +16,7 @@
 
 /* @flow */
 import { simulate } from 'glamor';
-import { mineralTheme } from '../../../../../../utils';
+import mineralTheme from '../../../../../../themes/mineral';
 import Menu, { MenuItem } from '../../../../../../Menu';
 import DemoLayout from '../../components/DemoLayout';
