+++ src/website/app/pages/ComponentDoc/index.js
@@ -16,7 +16,8 @@
 
 /* @flow */
 import React from 'react';
-import { createStyledComponent, mineralTheme } from '../../../../utils';
+import { createStyledComponent } from '../../../../utils';
+import mineralTheme from '../../../../themes/mineral';
 import Heading from '../../Heading';
 import DocBestPractices from './DocBestPractices';
 import DocExamples from './DocExamples';
